Update PDS4_CHAN1_IngestLDD.xml
A handful of the M3 L0 products have 'UNK' values in their spacecraft clock count fields. It might be reasonable to simply omit these fields from the PDS4 labels for these products, but at least one of these products is unusual in other ways (contrary to the target specified in the PDS3 label, it is an Earth-facing image), so we suspect that the 'UNK' literal may provide a good clue to end-users that a particular L0 product may not actually be part of the primary mapping sequence.